+++ clang-tools-extra/clang-tidy/hooks/TrivialCbakCheck.h
@@ -1,0 +1,35 @@
+//===--- TrivialCbakCheck.h - clang-tidy ------------------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_TRIVIALCBAKCHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_TRIVIALCBAKCHECK_H
+
+#include "../ClangTidyCheck.h"
+
+namespace clang {
+namespace tidy {
+namespace hooks {
+
+// For previous runtime versions, the 'cbak' entry point had been
+// mandatory, and therefore usually implemented with a body just
+// returning 0. Now this entry point is optional, it's better to save
+// space by not implementing it at all (unless needed for some
+// debugging scenarios).
+class TrivialCbakCheck : public ClangTidyCheck {
+public:
+  TrivialCbakCheck(StringRef Name, ClangTidyContext *Context)
+      : ClangTidyCheck(Name, Context) {}
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+};
+
+} // namespace hooks
+} // namespace tidy
+} // namespace clang
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_HOOKS_TRIVIALCBAKCHECK_H