+++ clang-tools-extra/test/clang-tidy/checkers/hooks-trivial-cbak.c
@@ -1,0 +1,9 @@
+// RUN: %check_clang_tidy %s hooks-trivial-cbak %t
+
+#include <stdint.h>
+
+// The parameter is missing but this check doesn't check that.
+int64_t cbak() {
+// CHECK-MESSAGES: :[[@LINE-1]]:1: warning: 'cbak' that just returns is not needed [hooks-trivial-cbak]
+    return 0;
+}